fix post reset pass controller
