+++ api/src/routes/resetpass.routes.js
@@ -1,5 +1,5 @@
 const routerResetPassword = require("express").Router();
-const handlerPostResetPassword = require("../handlers/resetPass/handlerPostResetPassw");
+const handlerPostResetPassword = require("../handlers/resetPass/handlerPostResetPassword");
 
 // Ruta para modificar contraseña y mandar correo 
 routerResetPassword.post("/", handlerPostResetPassword);
